requirements.txt
# local package
-e .

# external requirements
click~=6.7
Sphinx~=4.1.2

pandas~=1.3.3
param~=1.11.1
panel~=0.12.4
streamz~=0.6.3
numpy~=1.21.2
Pillow~=8.3.2
olefile~=0.45.1
pip~=21.2.4
cryptography~=2.1.4
tornado~=6.1
keyring~=10.6.0
toolz~=0.11.1
pyzmq~=22.2.1
colorama~=0.4.3
pexpect~=4.2.1
attrs~=21.2.0
packaging~=21.0
Jinja2~=3.0.1
PyYAML~=3.12
bokeh~=2.4.0
parso~=0.8.2
pyct~=0.4.8
requests~=2.18.4
pytz~=2018.3
HeapDict~=1.0.1
python-dateutil~=2.8.2
docutils~=0.15.2
Pygments~=2.10.0
setuptools~=57.5.0
six~=1.11.0
tqdm~=4.62.3
bleach~=4.1.0
jsonschema~=3.2.0
entrypoints~=0.3
webencodings~=0.5.1
MarkupSafe~=2.0.1
traitlets~=5.1.0
pygobject~=3.26.1
pickleshare~=0.7.5
wcwidth~=0.2.5
certifi~=2018.1.18
ptyprocess~=0.7.0
urllib3~=1.26.6
Markdown~=3.3.4
idna~=2.6
chardet~=3.0.4
simplejson~=3.13.2
pyparsing~=2.4.7
pycparser~=2.20
botocore~=1.21.36
pyrsistent~=0.18.0
zict~=2.0.0
zipp~=3.5.0